Refactor authentication and UI components
- Updated SplashActivity to use Application context for AuthRepository.
- Changed import path for ImageUtils in AddStockDialog.
- Modified EndShiftBottomSheet to use Double instead of BigDecimal for cash amounts.
- Adjusted IngredientListFragment to use updated method for user role retrieval.
- Refactored ShiftFragment to calculate current balance and updated cash handling methods.
- Updated StartShiftBottomSheet to validate opening cash as Double.
- Refined TableManagementFragment's search functionality with a TODO note for future implementation.
- Renamed methods in InventoryViewModel and ShiftViewModel for clarity.
- Enhanced LoginViewModel and RegisterViewModel to extend AndroidViewModel and utilize Application context.
- Cleaned up layout files for AdminHomeActivity and Ingredient Info Fragment.
- Removed unused Low Stock Badge from item_ingredient layout.
- Added new colors and strings for better UI consistency.
- Updated dependencies in libs.versions.toml for CircleImageView and core-i18n.
- Added a new profile icon drawable resource.

+++ app/src/main/java/com/group3/application/model/dto/EndShiftRequest.java
@@ -1,7 +1,5 @@
 package com.group3.application.model.dto;
 
-import java.math.BigDecimal;
-
 /**
  * Request DTO for ending a shift.
  * Matches the API endpoint: POST /shifts/end
@@ -9,21 +7,21 @@ import java.math.BigDecimal;
  */
 public class EndShiftRequest {
     
-    private BigDecimal closingCash;
+    private Double closingCash;
 
     // Constructors
     public EndShiftRequest() {}
 
-    public EndShiftRequest(BigDecimal closingCash) {
+    public EndShiftRequest(Double closingCash) {
         this.closingCash = closingCash;
     }
 
     // Getters and Setters
-    public BigDecimal getClosingCash() {
+    public Double getClosingCash() {
         return closingCash;
     }
 
-    public void setClosingCash(BigDecimal closingCash) {
+    public void setClosingCash(Double closingCash) {
         this.closingCash = closingCash;
     }
 }
